refactored tests for JaxRsTransactionNameInstrumentation
